refactor: Renamed resources to oresources
