+++ oresources/mock.go
@@ -1,4 +1,4 @@
-package resources
+package oresources
 
 import (
 	"encoding/json"
