+++ oresources/poresources.go
@@ -1,4 +1,4 @@
-package resources
+package oresources
 
 import (
 	"bytes"
